// src/server/models/enums.ts
export enum Allergies {
  Gluten = "Gluten",
  Lactose = "Lactose",
  Nuts = "Nuts",
  Peanuts = "Peanuts",
  Shellfish = "Shellfish",
  Fish = "Fish",
  Egg = "Egg",
  Soy = "Soy",
  Sesame = "Sesame",
  Celery = "Celery",
  Mustard = "Mustard",
  Molluscs = "Molluscs",
  Lupin = "Lupin",
  Sulfites = "Sulfites",
}

export enum FoodType {
  Unknown = "Unknown",
  Pork = "Pork",
  Chicken = "Chicken",
  Vegetarian = "Vegetarian",
  Vegan = "Vegan",
  Turkey = "Turkey",
  Beef = "Beef",
  Salat = "Salat",
  Fish = "Fish",
}